requirements.txt

flask
flask-sqlalchemy
flask-migrate
python-dotenv
pymysql